+++ src/losses/__init__.py
@@ -1,0 +1,3 @@
+from src.losses.segmentation import bce_dice_loss, dice_loss
+
+__all__ = ["dice_loss", "bce_dice_loss"]